+++ codex-rs/core/src/models_manager/manager.rs
