+++ codex-rs/core/src/models_manager/mod.rs
@@ -1,4 +1,4 @@
-mod cache;
+pub mod cache;
+pub mod manager;
 pub mod model_family;
 pub mod model_presets;
-pub mod models_manager;
